fix BitNum bug

+++ src/Common/Common.go
@@ -8,11 +8,11 @@ import (
 )
 
 func BitNum(num uint64) uint64 { //the number of bits needed by the num; 0 needs 0, 1 need 1, 2 need 2, 3 need 2 ....
-	var bitn uint64 = 63
-	for (bitn >= 0) && (((uint64(1) << uint32(bitn)) & num) == 0) {
+	var bitn int64 = 63
+	for (bitn >= 0) && (((1 << uint64(bitn)) & num) == 0) {
 		bitn--
 	}
-	return (bitn + 1)
+	return uint64(bitn + 1)
 }
 
 func Cmp(ai interface{}, bi interface{}) int {
